+++ packages/javascript-sdk/src/fr-auth/callbacks/ping-protect-initialize-callback.ts
@@ -3,7 +3,7 @@
  *
  * ping-protect-initialize-callback.ts
  *
- * Copyright (c) 2024 - 2025 Ping Identity Corporation. All rights reserved.
+ * Copyright (c) 2024 - 2026 Ping Identity Corporation. All rights reserved.
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
  */
@@ -26,6 +26,18 @@ class PingOneProtectInitializeCallback extends FRCallback {
    * Get callback's initialization config settings
    */
   public getConfig() {
+    const signalsInitializationOptions = this.getOutputByName<Record<string, unknown> | undefined>(
+      'signalsInitializationOptions',
+      undefined,
+    );
+    if (
+      signalsInitializationOptions !== null &&
+      typeof signalsInitializationOptions === 'object' &&
+      !Array.isArray(signalsInitializationOptions)
+    ) {
+      return signalsInitializationOptions;
+    }
+
     const agentIdentification = this.getOutputByName<boolean | undefined>(
       'agentIdentification',
       undefined,